requirements.txt
black==25.12.0
click==8.3.1
Jinja2==3.1.6
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pip==25.3
platformdirs==4.5.1
pytokens==0.3.0
PyYAML==6.0.3
ruff==0.14.8
setuptools==80.9.0
wheel==0.45.1
